ADFA-5598 | Move AI Code Suggestions to its own package, state the AI Core need
Rename the Java package to com.itsaky.androidide.plugins.aicodesuggestions and
follow it through namespace, applicationId, plugin.id and plugin.main_class.
getTooltipCategory() moves with plugin.id: the host derives the same string to
resolve a lookup, so a category that lags behind the id renders "n/a" at
runtime while the build stays green.

This changes plugin.id, so an existing install of the old id does not upgrade
to this one. The addon is on skip.txt and has never reached the gallery, so
there is nothing published to orphan.

State the AI Core requirement at the top of the README and the gallery page.
The plugin holds no model; it asks AI Core for every completion and shows
nothing until AI Core is installed and a model is selected.

Correct the build instructions: the addon carries its own Gradle wrapper, so
the commands are ./gradlew from plugins/AI-Code-Suggestions, and plugin-api.jar
resolves from ../../libs/.

Verified: ./gradlew assemblePlugin succeeds and packages both icons and
assets/docs/index.html. Not verified on a device.
