+++ plugins/AI-Code-Suggestions/README.md
@@ -1,13 +1,18 @@
 # AI Code Suggestions plugin for Code On The Go
 
+**Requires the AI Core addon.** This plugin holds no model of its own. It asks
+AI Core for every completion, so install AI Core first and select a model in
+**AI Assistant → AI Settings**. Without it, this plugin shows no completions.
+
 Inline **ghost-text** code completions. As you type, the plugin debounces, asks
 an LLM for a completion at the cursor, and shows it as dimmed inline text via
 `IdeEditorService.showInlineSuggestion()`.
 
-> This plugin has **no compile-time dependency** on `ai-core`; it resolves the
-> inference service at runtime through the Code On The Go plugin manager
-> (SharedServices). Install **`ai-core` first**, then this plugin, and configure
-> a model in **AI Assistant → AI Settings**.
+> The dependency is at **runtime only** — there is no compile-time dependency
+> on `ai-core`. This plugin resolves `LlmInferenceService` through the Code On
+> The Go plugin manager (SharedServices), and binds lazily. Until AI Core is
+> active it stays silent, then enables itself when the inference service
+> appears.
 
 ## Architecture
 
@@ -47,12 +52,12 @@ Prerequisites: Android SDK (API 33+), JDK 17. Create `local.properties` with
 `sdk.dir=...`. No NDK or native toolchain.
 
 ```bash
-cd AI-Code-Suggestions
-../gradlew assemblePlugin          # release  -> build/plugin/ai-code-suggestions.cgp
-../gradlew assemblePluginDebug     # debug variant
+cd plugins/AI-Code-Suggestions
+./gradlew assemblePlugin          # release  -> build/plugin/ai-code-suggestions.cgp
+./gradlew assemblePluginDebug     # debug variant
 ```
 
-The build resolves `plugin-api.jar` from the repo-root `../libs/`.
+The build resolves `plugin-api.jar` from the repo-root `../../libs/`.
 
 ## Installation
