+++ plugins/AI-Code-Suggestions/ai-code-suggestions.html
@@ -39,6 +39,13 @@
 <body>
   <h1>AI Code Suggestions Plugin</h1>
 
+  <div class="note">
+    <b>Requires the AI Core addon.</b> This plugin holds no model of its own —
+    it asks AI Core for every completion. Install <b>AI Core</b> first and
+    select a model in <b>AI Assistant &rarr; AI Settings</b>. Until you do, this
+    plugin shows no completions.
+  </div>
+
   <h2>Executive overview</h2>
   <p><b>AI Code Suggestions</b> adds inline <b>ghost-text</b> code completions to
   the Code On The Go editor. As you type, it asks the on-device model provided by
