+++ plugins/AI-Code-Suggestions/build.gradle.kts
@@ -9,11 +9,11 @@ pluginBuilder {
 }
 
 android {
-    namespace = "com.itsaky.androidide.plugins.codesuggestions"
+    namespace = "com.itsaky.androidide.plugins.aicodesuggestions"
     compileSdk = 36
 
     defaultConfig {
-        applicationId = "com.itsaky.androidide.plugins.codesuggestions"
+        applicationId = "com.itsaky.androidide.plugins.aicodesuggestions"
         minSdk = 33
         targetSdk = 36
         versionCode = 1
