+++ plugins/AI-Code-Suggestions/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.itsaky.androidide.plugins.codesuggestions">
+    package="com.itsaky.androidide.plugins.aicodesuggestions">
 
     <application
         android:label="AI Code Suggestions Plugin"
@@ -9,7 +9,7 @@
         <!-- Plugin metadata -->
         <meta-data
             android:name="plugin.id"
-            android:value="com.itsaky.androidide.plugins.codesuggestions" />
+            android:value="com.itsaky.androidide.plugins.aicodesuggestions" />
 
         <meta-data
             android:name="plugin.name"
@@ -37,7 +37,7 @@
 
         <meta-data
             android:name="plugin.main_class"
-            android:value="com.itsaky.androidide.plugins.codesuggestions.CodeSuggestionsPlugin" />
+            android:value="com.itsaky.androidide.plugins.aicodesuggestions.CodeSuggestionsPlugin" />
 
         <meta-data
             android:name="plugin.icon_day"
